requirements.txt
requests
python-dotenv
pillow
matplotlib
